Deletion Logic in, Element swapping rewritten
+ Deletion logic is working.  Buttons scheduled for deletion trigger a
check on their quadrant.
? Checking just rows would be more efficient, but I viewed the
investment in narrowing down the code as not worth it for the moment.
-/? There's a bug that (probably) only exists because I'm retaining
deleted squares for testing purposes.  However, it would be nice to find
the source and squash it.  The game is fun as is, and eliminating that
bug would open the door to putting an early version of it out to get
some feedback.

+ Element swapping is now done by swapping texture values instead of
swapping indexes (which negated the benefits of tying my data structure
to a physical structure!) and x/y coordinates.
? How inefficient is swapping textures?  Could it be more efficient by
using static texture declarations for the class?